AUI-1292 Unit test for aui-carousel module
Coverage of the test for the Carousel module should be 100% now. There were a few lines/branches pending but now all is covered.

While doing this I noticed that the listener for nodeMenuItemSelector changes wansn't working before. I updated its listener and made sure that registered the right event when changed.