AUI-1368 Modal can't be hidden if destroyOnHide is false
I noticed that if the destroyOnHide attribute is false on the current version of the modal, the modal is never hidden. That probably happened when switching to bootstrap3. Fixed the css to properly hide the modal.

I noticed then that when showing the modal again it wouldn't be properly aligned. That's because bootstrap uses relative position for .modal-dialog, but we need it to be fixed so we can spefify where it will show up. Fixed that as well.