AUI-1772 Fixing popover performance problem
Clicking on existing events was taking a long while to open popovers. I profiled scheduler and noticed this was happening because setting paths on YUI causes the contents to be cloned, which is taking a long while. I changed the call to instead set the main attribute with the desired values and this is not slow anymore.